LibMedia: Change the underlying type of CodecID to silence clang-tidy
We have warnings enabled for enums' underlying types being too large.

+++ Libraries/LibMedia/CodecID.h
@@ -11,7 +11,7 @@
 
 namespace Media {
 
-enum class CodecID : u32 {
+enum class CodecID : u8 {
     Unknown,
     // On2 / Google
     VP8,
